Move components and resuable functions
