+++ react/src/components/CategoricalTextOutput/index.tsx
