+++ react/src/components/CategoricalTextOutput/styles.scss
