+++ react/src/components/HoverDetails/index.tsx
