+++ react/src/components/HoverDetails/styles.scss
