+++ react/src/components/Information/EarthquakeAffectedLandlessHouseholds/index.tsx
