+++ react/src/components/Information/GeohazardAffectedHouseholds/DonutChart/Arc/index.tsx
