+++ react/src/components/Information/GeohazardAffectedHouseholds/DonutChart/Arc/styles.scss
