+++ react/src/components/Information/GeohazardAffectedHouseholds/DonutChart/index.tsx
