+++ react/src/components/Information/GeohazardAffectedHouseholds/index.tsx
