+++ react/src/components/Information/GeohazardAffectedHouseholds/styles.scss
