+++ react/src/components/Information/Header/styles.scss
