+++ react/src/components/Information/Heading/index.tsx
