+++ react/src/components/Information/Heading/styles.scss
