+++ react/src/components/Information/IntegratedSettlements/index.tsx
