+++ react/src/components/Information/LandPurchased/index.tsx
