+++ react/src/components/Information/LandslidesRiskScore/index.tsx
