+++ react/src/components/Information/LandslidesSurveyed/index.tsx
