+++ react/src/components/Information/PeopleRelocated/People/index.tsx
