+++ react/src/components/Information/PeopleRelocated/People/styles.scss
