+++ react/src/components/Information/PeopleRelocated/index.tsx
