+++ react/src/components/Information/PeopleRelocated/styles.scss
