+++ react/src/components/Information/styles.scss
