+++ react/src/components/RiskPointHoverDetails/index.tsx
