+++ react/src/components/RiskPointHoverDetails/styles.scss
