+++ react/src/components/TextOutput/index.tsx
