+++ react/src/components/TextOutput/styles.scss
